+++ src/LLInterface.cpp
@@ -81,7 +81,7 @@ std::map<std::string, double> LLInterface::GetAllSensors()
     sensors = HcpManager::GetAllSensors();
 
 #ifdef USE_TMPoE
-    std::vector<uint16> tmpValues = tmPoE->Read();
+    std::vector<uint32> tmpValues = tmPoE->Read();
     for (int i = 0; i < tmpValues.size(); i++)
     {
         sensors["temp " + std::to_string(i+1)] = tmpValues[i];
